requirements.txt
requests~=2.28.1
beautifulsoup4~=4.11.1
types-beautifulsoup4~=4.11.6.2
pytest-mock~=3.10.0
pytest~=7.2.0
psycopg2-binary~= 2.9.5
types-psycopg2~=2.9.21.2
mock~=5.0.0
flake8~=4.0.1
dataclasses-json~=0.5.7
python-dotenv~=0.21.0